[IMP] fastapi: Add fastapi addons directory structure into documentation
