+++ fastapi/tests/test_fastapi_demo.py
@@ -12,7 +12,7 @@ from odoo.tools import mute_logger
 from fastapi import status
 from fastapi.testclient import TestClient
 
-from .. import depends
+from .. import dependencies
 from ..context import odoo_env_ctx
 from ..models.fastapi_endpoint_demo import EndpointAppInfo, ExceptionType
 
@@ -32,7 +32,7 @@ class FastAPIDemoCase(TransactionCase):
         cls.test_partner = cls.env["res.partner"].create({"name": "FastAPI Demo"})
         cls.fastapi_demo_app = cls.env.ref("fastapi.fastapi_endpoint_demo")
         cls.app = cls.fastapi_demo_app._get_app()
-        cls.app.dependency_overrides[depends.authenticated_partner_impl] = partial(
+        cls.app.dependency_overrides[dependencies.authenticated_partner_impl] = partial(
             lambda a: a, cls.test_partner
         )
         # we need to disable the raise of unexpected exception into the called
